🚚 Rename and move the parser implementation to appropriate namespace
