+++ tests/MarkdownParserTest.php
@@ -2,7 +2,8 @@
 
 namespace Aheenam\Mozhi\Test;
 
-use Aheenam\Mozhi\MarkdownParser;
+
+use Aheenam\Mozhi\Documents\MarkdownDocument\CommonmarkParser;
 use Illuminate\Support\Facades\Config;
 use Webuni\CommonMark\TableExtension\TableExtension;
 
@@ -11,7 +12,7 @@ class MarkdownParserTest extends TestCase
     /** @test */
     public function it_can_parse_markdown_to_html()
     {
-        $parser = new MarkdownParser;
+        $parser = new CommonmarkParser;
 
         $this->assertEquals('<h1>Hello World</h1>', trim($parser->parse('# Hello World')));
     }
@@ -22,7 +23,7 @@ class MarkdownParserTest extends TestCase
         Config::set('markdown_extensions', [
             new TableExtension(),
         ]);
-        $parser = new MarkdownParser;
+        $parser = new CommonmarkParser;
 
         $this->assertTrue($parser->hasExtension(new TableExtension));
     }
